backend/requirements.txt
django==6.0.2
django-allauth==65.14.3
django-ninja==1.5.3
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
uvicorn==0.41.0
requests==2.32.5
django-ratelimit==4.1.0
python-ldap==3.4.5
pyyaml==6.0.3
amqp==5.3.1
celery==5.6.2